Refactor sanity lib usage
