+++ web/src/lib/sanity/sanity-api.ts
@@ -1,19 +1,5 @@
 import { client } from './client';
-import { PortableTextBlock, Slug } from '@sanity/types';
-import { Person } from '../types';
-
-export interface ManualSection {
-  title: string;
-  label: string;
-  slug: Slug;
-  content: Array<PortableTextBlock>;
-}
-
-export interface StaffManual {
-  title: string;
-  intro: string;
-  sections: Array<ManualSection>;
-}
+import { JobDescription, Person, StaffManual } from '../../types';
 
 export const getStaffManual = () =>
   client.fetch<StaffManual>(
@@ -25,14 +11,6 @@ export const getPeople = () =>
     '*[_type == "people" && !(_id in path(\'drafts.**\'))]{firstName, lastName, email, projects[] -> {name, slug}, socials, "imageUrl": image.asset->url}'
   );
 
-export interface JobDescription {
-  title: string;
-  slug: Slug;
-  label: string;
-  subtitle: string;
-  description: string;
-  sections: Array<ManualSection>;
-}
 export const getJobDescriptions = () =>
   client.fetch<JobDescription>(
     `*[_type == "job-description"][0]{ title, slug, label, subtitle, description}`
